fixed sql

+++ DDL/Tables/CreateTable-card.sql
@@ -4,14 +4,14 @@ DROP TABLE IF EXISTS card;
 
 CREATE TABLE card
 (
-    card_id      INT         NOT NULL AUTO_INCREMENT,
+    card_id      BIGINT      NOT NULL AUTO_INCREMENT,
     number       VARCHAR(20) NOT NULL,
-    payment_id   INT         NOT NULL,
+    payment_id   BIGINT      NOT NULL,
     start_date   DATE        NOT NULL,
-    card_type_id INT         NOT NULL,
+    card_type_id TINYINT     NOT NULL,
 
     CONSTRAINT PRIMARY KEY PK_card
-    (
+        (
          card_id
-    )
+            )
 );
