+++ DDL/Tables/CreateTable-order_status.sql
@@ -4,11 +4,11 @@ DROP TABLE IF EXISTS order_status;
 
 CREATE TABLE order_status
 (
-    status_id INT         NOT NULL AUTO_INCREMENT,
+    status_id TINYINT     NOT NULL AUTO_INCREMENT,
     status    VARCHAR(20) NOT NULL,
 
     CONSTRAINT PRIMARY KEY PK_order_status
-    (
+        (
          status_id
-    )
+            )
 );
